pyproject.toml
[project]
name = "suzuki-scraping"
version = "0.1.0"
description = ""
authors = [
    {name = "kai",email = "fkaiku42504261@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "pyautogui (>=0.9.54,<0.10.0)",
    "selenium (>=4.28.1,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
